tests/fixtures/requirements.txt
pyaml>=17.0,<18
pysha3>=1.0,<2.0
jsonschema>=2.6,<3.0
